pyproject.toml
[project]
name = "empart"
version = "0.1.0"
description = "Empart system is to transfer the medical records of a patient from one hospital to another with proper encryption using public key cryptography."
readme = "README.md"
authors = [
    { name = "Aero", email = "aerocyberdev@proton.me" }
]
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "flask>=3.1.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
]
